rename actor classes
